pyproject.toml
# --- PROJECT CONFIGURATION --- #

[build-system]
requires = ["setuptools>=67", "wheel", "setuptools-git-versioning>=2.0,<3"]
build-backend = "setuptools.build_meta"

# Metadata (see https://peps.python.org/pep-0621/)
[project]
name = "airflow-dags"
dynamic = ["version"] # Set automtically using git: https://setuptools-git-versioning.readthedocs.io/en/stable/
description = "Airflow DAGs for running Open Climate Fix's production pipelines"
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.12.0,<3.13"
license = {text = "MIT License"}
authors = [
    { name = "Open Climate Fix Team", email = "info@openclimatefix.org" },
    { name = "Peter Dudfield", email = "peter@openclimatefix.org" },
    { name = "Sol Cotton", email = "sol@openclimatefix.org" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "apache-airflow[amazon]==2.10.5",
    "apache-airflow-providers-slack==9.0.1",
    "xmlsec==1.3.14",
    "icechunk==0.2.17",
    "xarray==2025.6.1",
]

[dependency-groups]
dev = [
    # Testing
    "unittest-xml-reporting",
    # Linting and type checking
    "mypy",
    "ruff >= 0.9.2",
    "lxml-stubs",
    "types-Authlib",
    "types-Deprecated",
    "types-Pygments",
    "types-WTForms",
    "types-colorama",
    "types-croniter",
    "types-gevent",
    "types-jmespath",
    "types-jsonschema",
    "types-psutil",
    "types-python-dateutil",
    "types-requests",
    "types-tabulate",
    # LSP Support
    "python-lsp-server",
    "pylsp-mypy",
    "python-lsp-ruff",
    "types-pyyaml>=6.0.12.20250915",
]

[project.scripts]

[project.urls]
repository = "https://github.com/openclimatefix/airflow-dags"

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["test*"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.setuptools-git-versioning]
enabled = true

# --- LINTING AND TYPING CONFIGURATION --- #

# MyPy configuration
# * See https://mypy.readthedocs.io/en/stable/index.html
[tool.mypy]
python_version = "3.12"
strict = true
warn_unreachable = true
warn_return_any = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
# Ignore missing imports for libraries that don't have them.
# If they are ever made, remove from here!
module = [
    "botocore.session",
    "botocore.errorfactory",
    "botocore.client",
    "boto3",
    "fsspec",
    "fsspec.implementations",
    "fsspec.implementations.local",
    "fsspec.implementations.asyn_wrapper",
    "joblib",
    "s3fs",
    "xarray",
    "zarr",
]
ignore_missing_imports = true

# Ruff configuration
# * See https://beta.ruff.rs/docs/
[tool.ruff]
line-length = 100
indent-width = 4
exclude = ["__init__.py"]

[tool.ruff.lint]
select = [
    "F",   # pyflakes
    "E",   # pycodestyle
    "W",   # whitespace and newlines
    "I",   # isort
    "UP",  # modernize
    "ANN", # flake8 type annotations
    "S",   # flake8 bandit
    "B",   # flake8 bugbear
    "C4",  # flake8 comprehensions
    "COM", # flake8 commas
    "T20", # flake8 print
    "SIM", # flake8 simplify
    "ARG", # flake8 unused arguments
    "DTZ", # flake8 datetimes
    "Q",   # flake8 quotes
    "TCH", # flake8 typecheck
    "D",   # pydocstyle
    "RUF", # ruff-specific rules
]
fixable = ["ALL"]
ignore = [
    "D203", "D213", "D215", "D400", "D401", "D404", "D406",
    "D407", "D408", "D409", "D413",
]

[tool.ruff.lint.per-file-ignores]
"test*" = ["D", "ANN"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = 100

